ml_service/pyproject.toml
[project]
name = "ml-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.5.0",
    "faiss-gpu>=1.7.2",
    "fastapi>=0.128.0",
    "numpy>=1.26.4,<2.0.0",
    "pandas>=2.1.1,<2.2.0",
    "torchvision>=0.24.1",
    "transformers>=4.57.6",
    "uvicorn>=0.40.0",
]
